requirements.txt
viser
opencv-python
imageio
imageio-ffmpeg
matplotlib
tensorboard
scikit-learn
jaxtyping
roma
ninja
pytorch-msssim
fsspec
loguru
wandb
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.0
torchvision==0.17.0
transformers
einops
lpips
ipdb
nerfview==0.0.3
torchmetrics
splines==0.3.2
pyyaml
black==24.4.2
isort==5.13.2
--extra-index-url https://pypi.nvidia.com
cudf-cu11==24.6.*
dask-cudf-cu11==24.6.*
cuml-cu11==24.6.*
cugraph-cu11==24.6.*
cuspatial-cu11==24.6.*
cuproj-cu11==24.6.*
cuxfilter-cu11==24.6.*
cucim-cu11==24.6.*
pylibraft-cu11==24.6.*
raft-dask-cu11==24.6.*
cuvs-cu11==24.6.*
